SCUMM: Move high-level GUI handling to gfx_gui.cpp
